requirements.txt
pygame
PyOpenGL
numpy
PyGLM
Pillow